禁调高德和百度地图

+++ Baller/Classes/BallPark/BallerBallPark/Views/Baller_BallParkActivityListTableViewCell.m
@@ -41,7 +41,7 @@
     [self.imageView sd_setImageWithURL:[NSURL URLWithString:activitiyModel.user_photo] placeholderImage:[UIImage imageNamed:@"manHead"]];
     self.textLabel.text = activitiyModel.user_name;
     self.timeLabel.text = [TimeManager getMessageDateStringOfTimeInterval:activitiyModel.start_time];
-    self.menberLabel.text = $str(@"%ld人已加入",activitiyModel.join_num);
+    self.menberLabel.text = $str(@"%ld人已加入",(long)activitiyModel.join_num);
         
     if (activitiyModel.status == 1) {
         if ([TimeManager theSuccessivelyWithCurrentTimeFrom:activitiyModel.start_time]) {
